Move it all to the storage module
